Connected to MySQL database and integrated Address module for users. Implemented functionality to manage user addresses.

+++ sb-ecom/src/main/java/com/ecommerce/project/model/Role.java
@@ -16,7 +16,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "role_id")
-    private Integer roleId;
+    private long roleId;
 
     @ToString.Exclude
     @Enumerated(EnumType.STRING)
